+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+# Version (dev)
+* #573: Show filename in tooltip for custom screenshots.
+
+
 # Version 4.45.1
 * Fix the application looking in the wrong location for the compiler.